extract integration remote tests
