+++ transaction_test.go
@@ -25,7 +25,6 @@ import (
 
 	emulator "github.com/onflow/flow-emulator"
 	convert "github.com/onflow/flow-emulator/convert/sdk"
-	"github.com/onflow/flow-emulator/storage/remote"
 	"github.com/onflow/flow-emulator/types"
 )
 
@@ -1902,121 +1901,4 @@ func TestRollbackTransaction(t *testing.T) {
 
 	IncrementHelper(t, b, counterAddress, addTwoScript, 2)
 
-}
-
-func Test_SimulatedMainnetTransaction(t *testing.T) {
-	t.Parallel()
-	remoteStore, err := remote.New(remote.WithChainID(flowgo.Mainnet))
-	require.NoError(t, err)
-
-	b, err := emulator.NewBlockchain(
-		emulator.WithStore(remoteStore),
-		emulator.WithStorageLimitEnabled(false),
-		emulator.WithTransactionValidationEnabled(false),
-		emulator.WithChainID(flowgo.Mainnet),
-	)
-	require.NoError(t, err)
-
-	script := []byte(`
-		import Ping from 0x9799f28ff0453528
-		
-		transaction {
-			execute {
-				Ping.echo()
-			}
-		}
-	`)
-	addr := flowsdk.HexToAddress("0x9799f28ff0453528")
-	tx := flowsdk.NewTransaction().
-		SetScript(script).
-		SetGasLimit(flowgo.DefaultMaxTransactionGasLimit).
-		SetProposalKey(addr, 0, 0).
-		SetPayer(addr)
-
-	err = b.AddTransaction(*tx)
-	assert.NoError(t, err)
-
-	txRes, err := b.ExecuteNextTransaction()
-	assert.NoError(t, err)
-
-	_, err = b.CommitBlock()
-	assert.NoError(t, err)
-
-	//tx1Result, err := b.GetTransactionResult(tx.ID())
-	assert.NoError(t, err)
-	assert.NoError(t, txRes.Error)
-	assert.Len(t, txRes.Events, 1)
-	assert.Equal(t, txRes.Events[0].String(), "A.9799f28ff0453528.Ping.PingEmitted: 0x953f6f26d61710cb0e140bfde1022483b9ef410ddd181bac287d9968c84f4778")
-	assert.Equal(t, txRes.Events[0].Value.String(), `A.9799f28ff0453528.Ping.PingEmitted(sound: "ping ping ping")`)
-}
-
-func Test_SimulatedMainnetTransactionWithChanges(t *testing.T) {
-	t.Parallel()
-	remoteStore, err := remote.New(remote.WithChainID(flowgo.Mainnet))
-	require.NoError(t, err)
-
-	b, err := emulator.NewBlockchain(
-		emulator.WithStore(remoteStore),
-		emulator.WithStorageLimitEnabled(false),
-		emulator.WithTransactionValidationEnabled(false),
-		emulator.WithChainID(flowgo.Mainnet),
-	)
-	require.NoError(t, err)
-
-	script := []byte(`
-		import Ping from 0x9799f28ff0453528
-		
-		transaction {
-			execute {
-				Ping.sound = "pong pong pong"
-			}
-		}
-	`)
-	addr := flowsdk.HexToAddress("0x9799f28ff0453528")
-	tx := flowsdk.NewTransaction().
-		SetScript(script).
-		SetGasLimit(flowgo.DefaultMaxTransactionGasLimit).
-		SetProposalKey(addr, 0, 0).
-		SetPayer(addr)
-
-	err = b.AddTransaction(*tx)
-	assert.NoError(t, err)
-
-	txRes, err := b.ExecuteNextTransaction()
-	assert.NoError(t, err)
-	assert.NoError(t, txRes.Error)
-
-	_, err = b.CommitBlock()
-	assert.NoError(t, err)
-
-	script = []byte(`
-		import Ping from 0x9799f28ff0453528
-		
-		transaction {
-			execute {
-				Ping.echo()
-			}
-		}
-	`)
-	tx = flowsdk.NewTransaction().
-		SetScript(script).
-		SetGasLimit(flowgo.DefaultMaxTransactionGasLimit).
-		SetProposalKey(addr, 0, 0).
-		SetPayer(addr)
-
-	err = b.AddTransaction(*tx)
-	assert.NoError(t, err)
-
-	txRes, err = b.ExecuteNextTransaction()
-	assert.NoError(t, err)
-
-	_, err = b.CommitBlock()
-	assert.NoError(t, err)
-
-	//tx1Result, err := b.GetTransactionResult(tx.ID())
-	assert.NoError(t, err)
-	assert.NoError(t, txRes.Error)
-	assert.Len(t, txRes.Events, 1)
-	assert.Equal(t, txRes.Events[0].String(), "A.9799f28ff0453528.Ping.PingEmitted: 0x953f6f26d61710cb0e140bfde1022483b9ef410ddd181bac287d9968c84f4778")
-	assert.Equal(t, txRes.Events[0].Value.String(), `A.9799f28ff0453528.Ping.PingEmitted(sound: "pong pong pong")`)
 }
